Multithreading with errors

+++ src/infrastructure/peripherals/DistanceSensor.h
@@ -8,6 +8,8 @@
 #ifndef INFRASTRUCTURE_PERIPHERALS_DISTANCESENSOR_H_
 #define INFRASTRUCTURE_PERIPHERALS_DISTANCESENSOR_H_
 
+#include <stdint.h>
+
 namespace infrastructure
 {
 
@@ -17,7 +19,7 @@ class DistanceSensor
 public:
 
 	// Returns distance registered in centimeters
-	static float get_distance();
+	static int16_t get_distance();
 
 private:
